requirements.txt
langchain 
langchain-openai 
faiss-cpu 
tiktoken
unstructured
chromadb
python-dotenv
streamlit